+++ CHANGELOG.md
@@ -16,3 +16,4 @@ You can now take groups in your own direction by cloning the group index, while 
 - Fixes shortcut labels in editor formatting menu tooltips
 - Avoids duplicate Changes listed in version sidebar
 - Empty blocks at the end of draft are removed before saving
+- Fix "Trusted only" view of document, filtering changes from untrusted authors